moved functions and hardcoded inner pydantic models
